add:study backtracking
