+++ koreaygj/programmers/week2/studyW2P1_V2.cpp
