+++ workspace.code-workspace
@@ -1,0 +1,11 @@
+{
+  "folders": [
+	{
+		"path": "Level 1"
+	},
+	{
+		"path": "Level 2"
+	}
+],
+  "settings": {}
+}
